requirements.txt
anyio==3.6.2
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
fastapi==0.86.0
h11==0.14.0
idna==3.4
mongomock==4.1.2
packaging==21.3
pydantic==1.10.2
pymongo==4.3.2
pyparsing==3.0.9
python-dotenv==0.21.0
redis==4.3.4
sentinels==1.0.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
wrapt==1.14.1
